Add divided for components
