+++ src/components/ErrorComponent.tsx
@@ -1,0 +1,27 @@
+import classNames from 'classnames';
+import React from 'react';
+
+type Props = {
+  error: string;
+  onClose: () => void;
+};
+
+export const ErrorComponent: React.FC<Props> = ({ error, onClose }) => {
+  return (
+    <div
+      data-cy="ErrorNotification"
+      className={classNames(
+        'notification is-danger is-light has-text-weight-normal',
+        { hidden: !error },
+      )}
+    >
+      <button
+        data-cy="HideErrorButton"
+        type="button"
+        className="delete"
+        onClick={onClose}
+      />
+      {error}
+    </div>
+  );
+};
